src/requirements.txt
#av==8.0.2
# appdirs==1.4.4
# attrs==19.3.0
# beautifulsoup4==4.9.1
# certifi==2020.6.20
# Windows clipboard
clipboard~=0.0.4
# clipboard==0.0.4
colour==0.1.5
# EVE ESI
EsiPy==1.2.2
# importlib-metadata==1.7.0
cx_freeze~=6.4
# for CX-Freeze
packaging==20.4
# pyaml==20.4.0
# pyasn1==0.4.8
# pycodestyle==2.6.0
# pyflakes==2.2.0
# Sound
pyglet~=1.5.8
# all QT5
PyQt5~=5.15.1
PyQt5-sip~=12.8.1
PyQt5-stubs~=5.14.2
# PyQt5Designer==5.14.1
PyQtWebEngine~=5.15.1
# PyYAML==5.3.1
# URL calls
requests~=2.24.0
# rsa==4.6
# six==1.15.0
# soupsieve==2.0.1
# typed-ast==1.4.1
# validate-email==1.3
# zipp==3.4.0
#zroya==0.2.4

# HTML Tools
beautifulsoup4~=4.9.1
six~=1.15.0
# EsiPy Swagger
pyswagger~=0.8.39
# yaml~=0.2.5
pyyaml~=5.3.1